refactor(errcode): seal PublicDetail value as typed scalar funnel (#1035 review F1)
PR #1035 sealed PublicDetail/InternalDetail field shape but kept
PublicAttr(key, value any) — wire-unsafe values (chan, func, NaN/Inf,
map, struct, pointer) compiled fine and silently downgraded 4xx → 500
at json.Marshal time. Sealed external construction \!= sealed wire shape.

Close the value-type axis at the type system:

  - PublicDetail.value is now the sealed publicValue marker interface
    (unexported publicValue() method) — outside-package types cannot
    implement it.
  - Five typed concrete value wrappers: publicString / publicInt /
    publicBool / publicDuration / publicTime.
  - Five typed exported constructors replace PublicAttr(any):
    PublicString / PublicInt[T signed-integer] / PublicBool /
    PublicDuration / PublicTime. No PublicFloat (NaN/Inf risk) —
    callers strconv.FormatFloat to string at the call site.
  - InternalAttr keeps `any` value (server-only channel, no wire risk).

DETAILS-SEALED-FIELD-FROZEN-01 archtest (reflect-based field lock,
range pkg/errcode, reverse self-check via reflect.StructOf) guards the
sealed shape against in-package drift. PublicAttr / MustValidateDetailsKinds
gone; stale kernel/mustctor allowlist entry removed.

Mechanical migration of ~390 callsites across 95 production files
+ 2 active docs (gRPC adapter quickstart, capability-interaction-patterns)
+ ADR amendment threat-matrix re-eval. The pre-existing package-level
helper PublicString(err error) renamed to RenderPublic(err) to free
the slot for the new typed constructor — no external callers
(verified by repo grep).

AI-robust grading per .claude/rules/gocell/ai-robust.md:
  - Downstream Hard: typed Public* constructor surface enumerates
    wire-safe scalars; non-scalar Go types are inexpressible through
    any errcode entry point.
  - Upstream Hard: unexported {key, value} fields + sealed publicValue
    marker = no outside-package construction. Field reflect lock
    backstops in-package drift.

Refs: #1035 review F1/F2/F3/F4 findings (PublicDetail value-type
funnel + sealed field frozen archtest + stale Must* allowlist cleanup
+ active docs migration).
ref: ADR docs/architecture/202605051730-adr-errcode-message-pii-safety.md
     §Amendment 2026-05-27 (value-axis sealing rationale + threat
     matrix re-eval).

+++ adapters/postgres/outbox_writer_test.go
@@ -448,7 +448,7 @@ func TestOutboxWriter_WriteBatch_InvalidEntry(t *testing.T) {
 		assert.Contains(t, ecErrEmptyID.Message, "must not be empty")
 		attr, found := ecErrEmptyID.FindAttr("index")
 		require.True(t, found, "expected index detail")
-		assert.Equal(t, 1, attr.Value())
+		assert.Equal(t, int64(1), attr.Value())
 		assert.Empty(t, tx.execCalls)
 	})
 
@@ -464,7 +464,7 @@ func TestOutboxWriter_WriteBatch_InvalidEntry(t *testing.T) {
 		assert.Contains(t, ecErrZeroID.Message, "all-zeros")
 		attr, found := ecErrZeroID.FindAttr("index")
 		require.True(t, found, "expected index detail")
-		assert.Equal(t, 1, attr.Value())
+		assert.Equal(t, int64(1), attr.Value())
 		assert.Empty(t, tx.execCalls)
 	})
 }
